--- src/constants/bonuses.ts
@@ -1,48 +1,0 @@
-// https://github.com/SkyCryptWebsite/SkyCrypt/blob/development/common/constants/bonuses.js
-
-export const STATS_BONUS = {
-	// Skills
-	skill_farming: {
-		1: { health: 2, farming_fortune: 4 },
-		15: { health: 3, farming_fortune: 4 },
-		20: { health: 4, farming_fortune: 4 },
-		26: { health: 5, farming_fortune: 4 }
-	},
-	skill_mining: {
-		1: { defense: 1, mining_fortune: 4 },
-		15: { defense: 2, mining_fortune: 4 }
-	},
-	skill_combat: {
-		1: { crit_chance: 0.5 }
-	},
-	skill_foraging: {
-		1: { strength: 1, foraging_fortune: 4 },
-		15: { strength: 2, foraging_fortune: 4 }
-	},
-	skill_fishing: {
-		1: { health: 2 },
-		15: { health: 3 },
-		20: { health: 4 },
-		26: { health: 5 }
-	},
-	skill_enchanting: {
-		1: { intelligence: 1, ability_damage: 0.5 },
-		15: { intelligence: 2, ability_damage: 0.5 }
-	},
-	skill_alchemy: {
-		1: { intelligence: 1 },
-		15: { intelligence: 2 }
-	},
-	skill_taming: {
-		1: { pet_luck: 1 }
-	},
-	skill_dungeoneering: {
-		1: { health: 2 },
-		51: { health: 0 }
-	},
-	skill_social: {},
-	skill_carpentry: {
-		1: { health: 1 }
-	},
-	skill_runecrafting: {}
-};
